+++ ext/oj/oj.c
@@ -407,7 +407,7 @@ oj_parse_options(VALUE ropts, Options copts) {
 	    break;
 	case T_STRING:
 	    if (sizeof(copts->dump_opts.indent_str) <= (len = RSTRING_LEN(v))) {
-		rb_raise(rb_eArgError, "indent string is limited to %lu characters.", sizeof(copts->dump_opts.indent_str));
+		rb_raise(rb_eArgError, "indent string is limited to %lu characters.", (unsigned long)sizeof(copts->dump_opts.indent_str));
 	    }
 	    strcpy(copts->dump_opts.indent_str, StringValuePtr(v));
 	    copts->dump_opts.indent_size = (uint8_t)len;
@@ -558,7 +558,7 @@ oj_parse_options(VALUE ropts, Options copts) {
 	} else {
 	    rb_check_type(v, T_STRING);
 	    if (sizeof(copts->dump_opts.after_sep) <= (len = RSTRING_LEN(v))) {
-		rb_raise(rb_eArgError, "space string is limited to %lu characters.", sizeof(copts->dump_opts.after_sep));
+		rb_raise(rb_eArgError, "space string is limited to %lu characters.", (unsigned long)sizeof(copts->dump_opts.after_sep));
 	    }
 	    strcpy(copts->dump_opts.after_sep, StringValuePtr(v));
 	    copts->dump_opts.after_size = (uint8_t)len;
@@ -571,7 +571,7 @@ oj_parse_options(VALUE ropts, Options copts) {
 	} else {
 	    rb_check_type(v, T_STRING);
 	    if (sizeof(copts->dump_opts.before_sep) <= (len = RSTRING_LEN(v))) {
-		rb_raise(rb_eArgError, "sapce_before string is limited to %lu characters.", sizeof(copts->dump_opts.before_sep));
+		rb_raise(rb_eArgError, "sapce_before string is limited to %lu characters.", (unsigned long)sizeof(copts->dump_opts.before_sep));
 	    }
 	    strcpy(copts->dump_opts.before_sep, StringValuePtr(v));
 	    copts->dump_opts.before_size = (uint8_t)len;
@@ -584,7 +584,7 @@ oj_parse_options(VALUE ropts, Options copts) {
 	} else {
 	    rb_check_type(v, T_STRING);
 	    if (sizeof(copts->dump_opts.hash_nl) <= (len = RSTRING_LEN(v))) {
-		rb_raise(rb_eArgError, "object_nl string is limited to %lu characters.", sizeof(copts->dump_opts.hash_nl));
+		rb_raise(rb_eArgError, "object_nl string is limited to %lu characters.", (unsigned long)sizeof(copts->dump_opts.hash_nl));
 	    }
 	    strcpy(copts->dump_opts.hash_nl, StringValuePtr(v));
 	    copts->dump_opts.hash_size = (uint8_t)len;
@@ -597,7 +597,7 @@ oj_parse_options(VALUE ropts, Options copts) {
 	} else {
 	    rb_check_type(v, T_STRING);
 	    if (sizeof(copts->dump_opts.array_nl) <= (len = RSTRING_LEN(v))) {
-		rb_raise(rb_eArgError, "array_nl string is limited to %lu characters.", sizeof(copts->dump_opts.array_nl));
+		rb_raise(rb_eArgError, "array_nl string is limited to %lu characters.", (unsigned long)sizeof(copts->dump_opts.array_nl));
 	    }
 	    strcpy(copts->dump_opts.array_nl, StringValuePtr(v));
 	    copts->dump_opts.array_size = (uint8_t)len;
